Улучшает структуру проекта, реализовывает генерацию случайных дат и времени
