+++ src/components/events-list/trip-events/get-photos.js
@@ -1,0 +1,14 @@
+import {getRandomIntegerNumber} from '../../../helpers/utils.js';
+
+const getPhotoLink = () => {
+  return `http://picsum.photos/248/152?r=${Math.random()}`;
+};
+
+const getPhotos = () => {
+  const photosAmount = getRandomIntegerNumber(1, 5);
+  return new Array(photosAmount)
+    .fill(``)
+    .map(getPhotoLink);
+};
+
+export {getPhotos};
