+++ src/components/events-list/trip-events/get-random-description.js
@@ -1,0 +1,28 @@
+import {getRandomIntegerNumber} from '../../../helpers/utils.js';
+
+const eventDescriptions = [
+  `Lorem ipsum dolor sit amet, consectetur adipiscing elit. `,
+  `Cras aliquet varius magna, non porta ligula feugiat eget. `,
+  `Fusce tristique felis at fermentum pharetra. `,
+  `Aliquam id orci ut lectus varius viverra. `,
+  `Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. `,
+  `Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. `,
+  `Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. `,
+  `Sed sed nisi sed augue convallis suscipit in sed felis. `,
+  `Aliquam erat volutpat. `,
+  `Nunc fermentum tortor ac porta dapibus. `,
+  `In rutrum ac purus sit amet tempus. `
+];
+
+const getRandomDescription = () => {
+  const sentencesAmount = getRandomIntegerNumber(1, 5);
+  let description = ``;
+
+  for (let i = 0; i < sentencesAmount; i++) {
+    description += eventDescriptions[i];
+  }
+
+  return description.trim();
+};
+
+export {getRandomDescription};
