+++ src/components/trip-event-form/render-options.js
@@ -1,0 +1,8 @@
+const renderOptions = (cities) => {
+  return cities.map((city) => {
+    return (`<option value="${city}"></option>`);
+  })
+  .join(`\n`);
+};
+
+export {renderOptions};
