+++ src/components/trip-event-form/render-photos.js
@@ -1,0 +1,8 @@
+const renderPhotos = (photos) => {
+  return photos.map((photo) => {
+    return (`<img class="event__photo" src="${photo}" alt="Event photo">`);
+  })
+  .join(`\n`);
+};
+
+export {renderPhotos};
