+++ src/helpers/constants.js
@@ -1,0 +1,3 @@
+const TIME_PARSE_COEFFICIENT = 3600000;
+
+export {TIME_PARSE_COEFFICIENT};
